requirements.txt
asgiref==3.8.1
Django==5.1.4
sqlparse==0.5.3
psycopg[binary]==3.2.3
psycopg2==2.9.10
pillow==11.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1
requests_mock==1.12.1
django-cors-headers==4.6.0
django-rest-passwordreset==1.5.0